requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.26.5
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==20.1
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.1
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
yarl==1.8.2
